+++ api-js/src/organization/objects/__tests__/organization.test.js
@@ -15,7 +15,7 @@ import { databaseOptions } from '../../../../database-options'
 import { cleanseInput } from '../../../validators'
 import { Acronym, Slug } from '../../../scalars'
 import { domainLoaderConnectionsByOrgId } from '../../../domain/loaders'
-import { domainConnection } from '../../../domain/objects/domain'
+import { domainConnection } from '../../../domain/objects'
 import { affiliationConnectionLoaderByOrgId } from '../../../affiliation/loaders'
 import { affiliationConnection } from '../../../affiliation/objects'
 import { organizationType, organizationSummaryType } from '../../objects'
