+++ api-js/src/organization/objects/index.js
@@ -1,2 +1,3 @@
 export * from './organization'
+export * from './organization-connection'
 export * from './organization-summary'
